% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{A2barxn}
\alias{A2barxn}
\title{Second moment of continuous endowment insurance PV}
\usage{
A2barxn(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of second moments.
}
\description{
Computes \eqn{{}^{2}\bar{A}_{x:\angl{n}}}.
}
