% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{A2nEx}
\alias{A2nEx}
\title{Second moment of pure endowment PV}
\usage{
A2nEx(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of second moments.
}
\description{
Computes \eqn{{}^{2}{}_nE_x = (v')^n {}_n p_x}.
}
