% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{A2xn}
\alias{A2xn}
\title{Second moment of endowment insurance PV}
\usage{
A2xn(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of second moments.
}
\description{
Computes \eqn{{}^{2}A_{x:\angl{n}}}.
}
