% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_mthly.R
\name{A2xn1_m}
\alias{A2xn1_m}
\title{Second moment of m-thly term insurance PV}
\usage{
A2xn1_m(x, n, i, m, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of second moments.
}
\description{
Second moment of m-thly term insurance PV
}
