% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_mthly.R
\name{A2xn_m}
\alias{A2xn_m}
\title{Second moment of m-thly endowment insurance PV}
\usage{
A2xn_m(x, n, i, m, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of second moments.
}
\description{
Second moment of m-thly endowment insurance PV
}
