% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{Abarx_udd}
\alias{Abarx_udd}
\title{UDD approximation of continuous whole life insurance}
\usage{
Abarx_udd(Ax, i)
}
\arguments{
\item{Ax}{Discrete whole life insurance APV.}

\item{i}{Effective annual interest rate.}
}
\value{
Continuous whole life insurance APV under UDD.
}
\description{
Computes \eqn{\bar{A}_x = (i/\delta)A_x}.
}
