% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter14_multiple_decrement.R
\name{Abarxj_md}
\alias{Abarxj_md}
\title{Continuous multiple-decrement insurance present value}
\usage{
Abarxj_md(t, ptau, muj, delta, benefit = 1)
}
\arguments{
\item{t}{Strictly increasing nonnegative time points.}

\item{ptau}{In-force probabilities at the supplied time points.}

\item{muj}{Cause-specific decrement intensities at the supplied time points.}

\item{delta}{Force of interest. May be scalar or vector.}

\item{benefit}{Benefit payable on decrement. May be scalar or vector.}
}
\value{
A numeric vector of actuarial present values.
}
\description{
Approximates the actuarial present value of a benefit payable at the moment
of decrement using the trapezoidal rule.
}
\examples{
t <- seq(0, 20, by = 0.1)
ptau <- exp(-0.012 * t)
muj <- rep(0.002, length(t))
Abarxj_md(t, ptau, muj, delta = 0.05, benefit = 2000)

}
