% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{Abarxn}
\alias{Abarxn}
\title{Continuous endowment insurance APV}
\usage{
Abarxn(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of APVs.
}
\description{
Computes
\eqn{\bar{A}_{x:\angl{n}} = \bar{A}_{x:\angl{n}}^{1} + v^n\,{}_np_x}.
}
