% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{Abarxn1}
\alias{Abarxn1}
\title{Continuous term insurance APV}
\usage{
Abarxn1(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of APVs.
}
\description{
Computes
\eqn{\bar{A}_{x:\angl{n}}^{1} = \int_0^n v^t \, {}_t p_x \mu_{x+t}\,dt}.
}
