% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{Abarxn_udd}
\alias{Abarxn_udd}
\title{UDD approximation of continuous endowment insurance}
\usage{
Abarxn_udd(Axn1, nEx, i)
}
\arguments{
\item{Axn1}{Discrete term insurance APV.}

\item{nEx}{Pure endowment APV, \eqn{{}_nE_x}.}

\item{i}{Effective annual interest rate.}
}
\value{
Continuous endowment insurance APV under UDD.
}
\description{
Computes
\eqn{\bar{A}_{x:\overline{n}|} = (i/\delta)A_{x:\overline{n}|}^{1} + {}_nE_x}.
}
