% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{Abarxy}
\alias{Abarxy}
\title{Continuous joint-life whole life insurance}
\usage{
Abarxy(x, y, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\overline{A}_{xy} = 1 - \delta \overline{a}_{xy}}.
}
\examples{
Abarxy(40, 50, i = 0.05, model = "uniform", omega = 100)
}
