% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{Abarxy1}
\alias{Abarxy1}
\title{Continuous contingent insurance: benefit on death of (x) if before (y)}
\usage{
Abarxy1(x, y, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\overline{A}_{xy}^{1} = \int_0^\infty v^t\,{}_tp_{xy}\mu_{x+t}\,dt}.
}
\examples{
Abarxy1(40, 50, i = 0.05, model = "uniform", omega = 100)
}
