% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{Abarxybar}
\alias{Abarxybar}
\title{Continuous last-survivor whole life insurance}
\usage{
Abarxybar(x, y, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\overline{A}_{\overline{xy}} = \overline{A}_x + \overline{A}_y - \overline{A}_{xy}}.
}
\examples{
Abarxybar(40, 50, i = 0.05, model = "uniform", omega = 100)
}
