% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{abarxybar_ch12}
\alias{abarxybar_ch12}
\alias{abarxybar}
\title{Continuous last-survivor whole life annuity}
\usage{
abarxybar(x, y, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\overline{a}_{\overline{xy}} = \overline{a}_x + \overline{a}_y - \overline{a}_{xy}}.
}
\details{
Shared documentation topic used to avoid filename collisions with
case-distinct function names on case-insensitive file systems.
}
\examples{
abarxybar(40, 50, i = 0.05, model = "uniform", omega = 100)
}
