% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{abary_x}
\alias{abary_x}
\title{Continuous reversionary annuity to (x) after death of (y)}
\usage{
abary_x(x, y, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\overline{a}_{y|x} = \overline{a}_x - \overline{a}_{xy}}.
}
\examples{
abary_x(40, 50, i = 0.05, model = "uniform", omega = 100)
}
