% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{adotxyn}
\alias{adotxyn}
\title{Joint-life temporary annuity-due}
\usage{
adotxyn(x, y, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\ddot{a}_{xy:\angl{n}}}.
}
\examples{
adotxyn(40, 50, n = 10, i = 0.05, model = "uniform", omega = 100)
}
