% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_approximations.R
\name{annuity_approximations}
\alias{annuity_approximations}
\title{Annuity approximations (Chapter 8)}
\description{
Chapter 8 approximation formulas for m-thly and continuous life annuities.
}
\details{
This file implements:
\itemize{
  \item UDD approximations for m-thly annuities,
  \item UDD approximations for continuous annuities,
  \item Woolhouse 2-term approximations,
  \item Woolhouse 3-term approximations.
}
}
