% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_approximations.R
\name{annuity_approximations_woolhouse2}
\alias{annuity_approximations_woolhouse2}
\alias{ax_m_woolhouse2}
\alias{adotx_m_woolhouse2}
\alias{axn_m_woolhouse2}
\alias{adotxn_m_woolhouse2}
\alias{nax_m_woolhouse2}
\alias{nadotx_m_woolhouse2}
\alias{sxn_m_woolhouse2}
\alias{sdotxn_m_woolhouse2}
\alias{abarx_woolhouse2}
\title{Woolhouse 2-term annuity approximations}
\usage{
ax_m_woolhouse2(x, m, i, model, ...)

adotx_m_woolhouse2(x, m, i, model, ...)

nax_m_woolhouse2(x, n, m, i, model, ...)

nadotx_m_woolhouse2(x, n, m, i, model, ...)

axn_m_woolhouse2(x, n, m, i, model, ...)

adotxn_m_woolhouse2(x, n, m, i, model, ...)

sxn_m_woolhouse2(x, n, m, i, model, ...)

sdotxn_m_woolhouse2(x, n, m, i, model, ...)

abarx_woolhouse2(x, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{m}{Number of payments per year.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}

\item{n}{Term.}
}
\value{
Numeric vector.
}
\description{
Woolhouse 2-term approximations for Chapter 8 annuity functions.
}
