% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_approximations.R
\name{annuity_approximations_woolhouse3}
\alias{annuity_approximations_woolhouse3}
\alias{ax_m_woolhouse3}
\alias{adotx_m_woolhouse3}
\alias{axn_m_woolhouse3}
\alias{adotxn_m_woolhouse3}
\alias{nax_m_woolhouse3}
\alias{nadotx_m_woolhouse3}
\alias{abarx_woolhouse3}
\title{Woolhouse 3-term annuity approximations}
\usage{
ax_m_woolhouse3(x, m, i, model, ...)

adotx_m_woolhouse3(x, m, i, model, ...)

nax_m_woolhouse3(x, n, m, i, model, ...)

nadotx_m_woolhouse3(x, n, m, i, model, ...)

axn_m_woolhouse3(x, n, m, i, model, ...)

adotxn_m_woolhouse3(x, n, m, i, model, ...)

abarx_woolhouse3(x, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{m}{Number of payments per year.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}

\item{n}{Term.}
}
\value{
Numeric vector.
}
\description{
Woolhouse 3-term approximations for Chapter 8 annuity functions.
}
