% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_mthly.R
\name{annuity_mthly}
\alias{annuity_mthly}
\alias{ax_m}
\alias{adotx_m}
\alias{axn_m}
\alias{adotxn_m}
\alias{nax_m}
\alias{nadotx_m}
\alias{sxn_m}
\alias{sdotxn_m}
\title{m-thly contingent annuity functions}
\usage{
ax_m(x, m, i, model, ..., k_max = 2e+05, tol = 1e-12)

adotx_m(x, m, i, model, ..., k_max = 2e+05, tol = 1e-12)

axn_m(x, n, m, i, model, ...)

adotxn_m(x, n, m, i, model, ...)

nax_m(x, n, m, i, model, ..., k_max = 2e+05, tol = 1e-12)

nadotx_m(x, n, m, i, model, ..., k_max = 2e+05, tol = 1e-12)

sxn_m(x, n, m, i, model, ...)

sdotxn_m(x, n, m, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{m}{Number of payments per year.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model name.}

\item{...}{Additional parameters passed to the survival model.}

\item{k_max}{Maximum summation horizon for non-terminating models.}

\item{tol}{Truncation tolerance for non-terminating models.}

\item{n}{Term or deferral period in years.}
}
\value{
Numeric vector containing the requested m-thly annuity value or actuarial
accumulated value.
}
\description{
Life annuities payable m-thly.
}
