% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_mthly.R
\name{annuity_mthly_accum_immediate}
\alias{annuity_mthly_accum_immediate}
\alias{sxn_m}
\title{Temporary m-thly annuity-immediate actuarial accumulated value}
\usage{
sxn_m(x, n, m, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term in years.}

\item{m}{Number of payments per year.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model name.}

\item{...}{Additional parameters passed to the survival model.}
}
\value{
Numeric vector of actuarial accumulated values.
}
\description{
Computes
\deqn{s_{x:\angl{n}}^{(m)} = a_{x:\angl{n}}^{(m)} / {}_nE_x}
}
\examples{
sxn_m(40, n = 10, m = 12, i = 0.05, model = "uniform", omega = 100)
}
