% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_mthly.R
\name{annuity_mthly_temp_due}
\alias{annuity_mthly_temp_due}
\alias{adotxn_m}
\title{Temporary m-thly annuity-due}
\usage{
adotxn_m(x, n, m, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term in years.}

\item{m}{Number of payments per year.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model name.}

\item{...}{Additional parameters passed to the survival model.}
}
\value{
Numeric vector of annuity values.
}
\description{
Computes the exact temporary m-thly annuity-due.
}
\details{
\deqn{\ddot{a}_{x:\angl{n}}^{(m)} = \frac{1}{m}\sum_{t=0}^{mn-1} v^{t/m}\,{}_{t/m}p_x}
}
\examples{
adotxn_m(40, n = 10, m = 12, i = 0.05, model = "uniform", omega = 100)
}
