% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{APV_gross_premiums}
\alias{APV_gross_premiums}
\title{APV of gross premiums under a risk discount rate}
\usage{
APV_gross_premiums(G, r, p_tau)
}
\arguments{
\item{G}{Gross premium vector for policy years 1 through \eqn{n}.}

\item{r}{Risk discount rate.}

\item{p_tau}{One-year in-force probabilities. This may have length
\eqn{n-1} or \eqn{n}. If length \eqn{n}, the final entry is ignored.}
}
\value{
Numeric scalar.
}
\description{
Computes the actuarial present value of gross premiums:
\deqn{
APV_{GP} = \sum_{t=0}^{n-1} \frac{G_{t+1} \cdot {}_tp_x^{(\tau)}}{(1+r)^t}.
}
}
\examples{
APV_gross_premiums(G = rep(95, 3),r = 0.10,p_tau = c(0.99858, 0.99847, 0.99834))

}
