% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{Ax_m_udd}
\alias{Ax_m_udd}
\title{UDD approximation of m-thly whole life insurance}
\usage{
Ax_m_udd(Ax, i, m)
}
\arguments{
\item{Ax}{Discrete whole life insurance APV.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}
}
\value{
m-thly whole life insurance APV under UDD.
}
\description{
Computes \eqn{A_x^{(m)} = (i/i^{(m)})A_x}.
}
