% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{Axn1}
\alias{Axn1}
\title{Term insurance APV}
\usage{
Axn1(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of APVs.
}
\description{
Computes \eqn{A_{x:\angl{n}}^{1} = \sum_{k=0}^{n-1} v^{k+1} {}_{k\mid}q_x}.
}
