% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_mthly.R
\name{Axn1_m}
\alias{Axn1_m}
\title{m-thly term insurance APV}
\usage{
Axn1_m(x, n, i, m, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of APVs.
}
\description{
Computes
\eqn{A_{x:\angl{n}}^{1(m)} = \sum_{j=0}^{mn-1} v^{(j+1)/m}\Pr(j/m < T_x \le (j+1)/m)}.
}
