% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{Axn1_m_udd}
\alias{Axn1_m_udd}
\title{UDD approximation of m-thly term insurance}
\usage{
Axn1_m_udd(Axn1, i, m)
}
\arguments{
\item{Axn1}{Discrete term insurance APV.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}
}
\value{
m-thly term insurance APV under UDD.
}
\description{
Computes \eqn{A_{x:\angl{n}}^{1(m)} = (i/i^{(m)})A_{x:\angl{n}}^{1}}.
}
