% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_mthly.R
\name{Axn_m}
\alias{Axn_m}
\title{m-thly endowment insurance APV}
\usage{
Axn_m(x, n, i, m, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of APVs.
}
\description{
Computes
\eqn{A_{x:\angl{n}}^{(m)} = A_{x:\angl{n}}^{1(m)} + v^n\,{}_np_x}.
}
