% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{Axn_m_udd}
\alias{Axn_m_udd}
\title{UDD approximation of m-thly endowment insurance}
\usage{
Axn_m_udd(Axn1, nEx, i, m)
}
\arguments{
\item{Axn1}{Discrete term insurance APV.}

\item{nEx}{Pure endowment APV.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}
}
\value{
m-thly endowment insurance APV under UDD.
}
\description{
Computes
\eqn{A_{x:\angl{n}}^{(m)} = (i/i^{(m)})A_{x:\angl{n}}^{1} + {}_nE_x}.
}
