% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{Axybarn}
\alias{Axybarn}
\title{Last-survivor endowment insurance}
\usage{
Axybarn(x, y, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{A_{\overline{xy}:\angl{n}}}.
}
\examples{
Axybarn(40, 50, n = 10, i = 0.05, model = "uniform", omega = 100)
}
