% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{Axyn}
\alias{Axyn}
\title{Joint-life endowment insurance}
\usage{
Axyn(x, y, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{A_{xy:\overline{n}|}}.
}
\examples{
Axyn(40, 50, n = 10, i = 0.05, model = "uniform", omega = 100)
}
