% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{ch12_multilife_cont_params}
\alias{ch12_multilife_cont_params}
\title{Shared parameters for Chapter 12 continuous multi-life functions}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{i}{Effective annual interest rate.}

\item{n}{Term in years.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\description{
Shared parameters for Chapter 12 continuous multi-life functions
}
\keyword{internal}
