% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{cov_term_deferred}
\alias{cov_term_deferred}
\title{Covariance of term and deferred insurance PVs}
\usage{
cov_term_deferred(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of covariances.
}
\description{
Computes
\eqn{\mathrm{Cov}(Z_{x:\angl{n}}^{1}, {}_{n\mid}Z_x) = -A_{x:\angl{n}}^{1} \cdot {}_{n\mid}A_x}.
}
