% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{cov_term_endow}
\alias{cov_term_endow}
\title{Covariance of term insurance and pure endowment PVs}
\usage{
cov_term_endow(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of covariances.
}
\description{
Computes
\eqn{\mathrm{Cov}(Z_{x:\angl{n}}^{1}, Z_{x:\angl{n}}^{1\text{(pure endow)}}) =
-A_{x:\angl{n}}^{1} \cdot {}_nE_x}.
}
