% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_models.R
\name{cumhaz0}
\alias{cumhaz0}
\title{Cumulative hazard for age-at-failure}
\usage{
cumhaz0(t, model, ...)
}
\arguments{
\item{t}{Numeric vector of times.}

\item{model}{One of \code{"uniform"}, \code{"exponential"},
\code{"gompertz"}, \code{"makeham"}, or \code{"weibull"}.}

\item{...}{Model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\Lambda_0(t)}.
}
