% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_varying_benefits.R
\name{DAbarxn1}
\alias{DAbarxn1}
\title{Piecewise-continuous decreasing n-year term insurance}
\usage{
DAbarxn1(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes
\deqn{(D\bar{A})_{x:\angl{n}}^{1} = \int_0^n \lfloor n+1-t \rfloor\, v^t\, {}_tp_x\, \mu_{x+t}\, dt.}
}
