% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_theory.R
\name{discount}
\alias{discount}
\title{Discount factor for compound interest}
\usage{
discount(i, t)
}
\arguments{
\item{i}{Effective interest rate.}

\item{t}{Time (can be vector).}
}
\value{
Discount factor at time t.
}
\description{
Discount factor for compound interest
}
\examples{
discount(0.05, 0:5)
}
