% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_models.R
\name{dist0}
\alias{dist0}
\alias{F0}
\alias{f0}
\title{Distribution functions for age-at-failure}
\usage{
F0(t, model, ...)

f0(t, model, ...)
}
\arguments{
\item{t}{Numeric vector of times.}

\item{model}{One of \code{"uniform"}, \code{"exponential"},
\code{"gompertz"}, \code{"makeham"}, or \code{"weibull"}.}

\item{...}{Model parameters.}
}
\value{
Numeric vector.
}
\description{
Distribution functions for age-at-failure
}
