% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{double_force_delta}
\alias{double_force_delta}
\title{Doubled force of interest}
\usage{
double_force_delta(delta)
}
\arguments{
\item{delta}{Numeric vector of forces of interest.}
}
\value{
Numeric vector of doubled forces of interest.
}
\description{
Computes \eqn{\delta' = 2\delta}.
}
