% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{double_force_i}
\alias{double_force_i}
\title{Effective annual interest at doubled force}
\usage{
double_force_i(i)
}
\arguments{
\item{i}{Numeric vector of effective annual interest rates.}
}
\value{
Numeric vector of effective annual rates corresponding to doubled force.
}
\description{
If \eqn{\delta' = 2\delta}, then
\eqn{i' = (1+i)^2 - 1}.
}
