% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{dx}
\alias{dx}
\title{Compute deaths between ages x and x+1}
\usage{
dx(tbl, x)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}
}
\value{
Numeric vector of d_x values.
}
\description{
Compute deaths between ages x and x+1
}
