% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{dxj}
\alias{dxj}
\title{Cause-specific decrements d_x^(j)}
\usage{
dxj(lxtau, qxj)
}
\arguments{
\item{lxtau}{Number alive at age x in the multiple-decrement table.}

\item{qxj}{Numeric vector of cause-specific decrement probabilities.}
}
\value{
Numeric vector.
}
\description{
Cause-specific decrements d_x^(j)
}
\examples{
dxj(1000, c(0.011, 0.100))
}
