% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_expectations.R
\name{ex_complete_tab}
\alias{ex_complete_tab}
\title{Complete expectation of life from a life table}
\usage{
ex_complete_tab(tbl, x, assumption = c("udd", "cf", "balducci"))
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of integer ages.}

\item{assumption}{One of \code{"udd"}, \code{"cf"}, \code{"balducci"}.}
}
\value{
Numeric vector of complete expectations \eqn{\overset{\circ}{e}_x}.
}
\description{
Computes
\eqn{\overset{\circ}{e}_x = \int_0^\infty {}_t p_x \, dt}
using a within-year assumption:
\code{"udd"}, \code{"cf"}, or \code{"balducci"}.
}
