% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_expectations.R
\name{ex_curtate_tab}
\alias{ex_curtate_tab}
\title{Curtate expectation of life from a life table}
\usage{
ex_curtate_tab(tbl, x)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of integer ages.}
}
\value{
Numeric vector of curtate expectations \eqn{e_x}.
}
\description{
Computes the curtate expectation of life
\eqn{e_x = \sum_{k=1}^\infty {}_k p_x}
in the discrete tabular setting.
}
