% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_expectations.R
\name{ex_temp_complete_tab}
\alias{ex_temp_complete_tab}
\title{Temporary complete expectation of life from a life table}
\usage{
ex_temp_complete_tab(tbl, x, n, assumption = c("udd", "cf", "balducci"))
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of integer ages.}

\item{n}{Numeric vector of nonnegative numbers.}

\item{assumption}{One of \code{"udd"}, \code{"cf"}, \code{"balducci"}.}
}
\value{
Numeric vector of temporary complete expectations.
}
\description{
Computes
\eqn{\overset{\circ}{e}_{x:\angl{n}} = \int_0^n {}_t p_x \, dt}
using a within-year assumption:
\code{"udd"}, \code{"cf"}, or \code{"balducci"}.
}
