% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_expectations.R
\name{ex_temp_curtate_tab}
\alias{ex_temp_curtate_tab}
\title{Temporary curtate expectation of life from a life table}
\usage{
ex_temp_curtate_tab(tbl, x, n)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of integer ages.}

\item{n}{Numeric vector of nonnegative integers.}
}
\value{
Numeric vector of temporary curtate expectations.
}
\description{
Computes
\eqn{e_{x:\angl{n}} = \sum_{k=1}^{n} {}_k p_x}
for integer n in the discrete tabular setting.
}
