% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_ch10.R
\name{GI_disc}
\alias{GI_disc}
\title{Interest gain for a discrete insurance contract}
\usage{
GI_disc(Vt, Vt1, P, i_actual, q_assumed, B = 1)
}
\arguments{
\item{Vt}{Reserve at duration t.}

\item{Vt1}{Reserve at duration t+1.}

\item{P}{Net premium for the year.}

\item{i_actual}{Actual annual effective interest rate.}

\item{q_assumed}{Assumed mortality rate for the year.}

\item{B}{Benefit amount. Defaults to 1.}
}
\value{
A numeric vector of values.
}
\description{
Interest gain for a discrete insurance contract
}
\examples{
GI_disc(Vt = 0.1, Vt1 = 0.11, P = 0.02, i_actual = 0.05, q_assumed = 0.01)
}
