% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_ch10.R
\name{GM_disc}
\alias{GM_disc}
\title{Mortality gain for a discrete insurance contract}
\usage{
GM_disc(Vt, Vt1, P, i_assumed, q_actual, B = 1)
}
\arguments{
\item{Vt}{Reserve at duration t.}

\item{Vt1}{Reserve at duration t+1.}

\item{P}{Net premium for the year.}

\item{i_assumed}{Assumed annual effective interest rate.}

\item{q_actual}{Actual mortality rate for the year.}

\item{B}{Benefit amount. Defaults to 1.}
}
\value{
Numeric vector.
}
\description{
Mortality gain for a discrete insurance contract
}
\examples{
GM_disc(Vt = 0.1, Vt1 = 0.11, P = 0.02, i_assumed = 0.04, q_actual = 0.01)
}
