% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_models.R
\name{hazard0}
\alias{hazard0}
\title{Hazard or force of mortality for age-at-failure}
\usage{
hazard0(t, model, ...)
}
\arguments{
\item{t}{Numeric vector of times.}

\item{model}{One of \code{"uniform"}, \code{"exponential"},
\code{"gompertz"}, \code{"makeham"}, or \code{"weibull"}.}

\item{...}{Model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{\lambda_0(t)}.
}
