% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_varying_benefits.R
\name{IAbarx}
\alias{IAbarx}
\title{Piecewise-continuous increasing whole life insurance}
\usage{
IAbarx(x, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes
\deqn{(I\bar{A})_x = \int_0^\infty \lfloor t+1 \rfloor\, v^t\, {}_tp_x\, \mu_{x+t}\, dt.}
}
