% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_varying_benefits.R
\name{IbarAbarx}
\alias{IbarAbarx}
\title{Fully continuous increasing whole life insurance}
\usage{
IbarAbarx(x, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes
\deqn{(\bar{I}\bar{A})_x = \int_0^\infty t\, v^t\, {}_tp_x\, \mu_{x+t}\, dt.}
}
