% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_varying_benefits.R
\name{IbarAbarxn1}
\alias{IbarAbarxn1}
\title{Fully continuous increasing n-year term insurance}
\usage{
IbarAbarxn1(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes
\deqn{(\bar{I}\bar{A})_{x:\angl{n}}^{1} = \int_0^n t\, v^t\, {}_tp_x\, \mu_{x+t}\, dt.}
}
