% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{Income_dc}
\alias{Income_dc}
\title{Retirement income from a defined contribution accumulation}
\usage{
Income_dc(AVz, adue_z)
}
\arguments{
\item{AVz}{Nonnegative accumulated value at retirement.}

\item{adue_z}{Positive whole-life annuity-due factor at retirement.}
}
\value{
A numeric vector.
}
\description{
Converts an accumulated account value into annual annuity-due income.
Scalar arguments are recycled to a common length.
}
\examples{
Income_dc(AVz = 824211.35, adue_z = 12)

}
