% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_varying_benefits.R
\name{insurance_varying_benefits}
\alias{insurance_varying_benefits}
\title{Insurance models with varying benefits}
\description{
Functions for discrete and continuous contingent payment models with
varying benefits.
}
\keyword{internal}
