% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{IRR_profit}
\alias{IRR_profit}
\title{Internal rate of return of a profit signature}
\usage{
IRR_profit(Pi, interval = c(0, 1), tol = .Machine$double.eps^0.5)
}
\arguments{
\item{Pi}{Profit signature vector.}

\item{interval}{Numeric vector of length 2 giving the search interval for
\code{uniroot()}.}

\item{tol}{Tolerance passed to \code{uniroot()}.}
}
\value{
Numeric scalar.
}
\description{
Computes the internal rate of return (IRR), defined as the rate \eqn{r} for
which the net present value is zero.
}
\examples{
Pi <- c(-15.00, 8.42, 8.39, 8.58)
IRR_profit(Pi)

}
