% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{last_survivor_annuities}
\alias{last_survivor_annuities}
\alias{adotxybarn}
\alias{axybarn}
\alias{adotxybar}
\alias{axybar}
\title{Last-survivor annuity functions}
\usage{
adotxybarn(x, y, n, i, tbl = NULL, model = NULL, ...)

axybarn(x, y, n, i, tbl = NULL, model = NULL, ...)

adotxybar(x, y, i, tbl = NULL, model = NULL, ..., k_max = 5000L, tol = 1e-12)

axybar(x, y, i, tbl = NULL, model = NULL, ..., k_max = 5000L, tol = 1e-12)
}
\arguments{
\item{x}{Age of the first life. May be scalar or vector.}

\item{y}{Age of the second life. May be scalar or vector.}

\item{n}{Term in years. May be scalar or vector.}

\item{i}{Effective annual interest rate. May be scalar or vector.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model.}

\item{...}{Additional parameters passed to the survival model or
life-table functions.}

\item{k_max}{Maximum number of terms used for an infinite series.}

\item{tol}{Numerical tolerance used to assess convergence.}
}
\value{
A numeric vector of actuarial present values.
}
\description{
Computes temporary and whole-life annuities for two lives under the
last-survivor status.
}
