% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{lx}
\alias{lx}
\title{Extract life-table survivor values}
\usage{
lx(tbl, x)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}
}
\value{
Numeric vector of l_x values.
}
\description{
Extract life-table survivor values
}
