% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_select.R
\name{lx_select}
\alias{lx_select}
\title{Extract select-table survivor value}
\usage{
lx_select(tbl, x_sel, t)
}
\arguments{
\item{tbl}{A select_life_table object.}

\item{x_sel}{Numeric vector of ages at selection.}

\item{t}{Numeric vector of durations since selection.}
}
\value{
Numeric vector of survivor values.
}
\description{
Returns \eqn{l_{[x]+t}} from a select life table.
}
