% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_conversions.R
\name{lx_to_S0}
\alias{lx_to_S0}
\title{Convert life-table values to survival probabilities}
\usage{
lx_to_S0(lx)
}
\arguments{
\item{lx}{Numeric vector of life-table survivor values.}
}
\value{
Numeric vector of \eqn{S_0(x)} values.
}
\description{
Converts Chapter 6 life-table values \eqn{l_x} into Chapter 5
survival probabilities \eqn{S_0(x) = l_x / l_0}.
}
