% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{md_table}
\alias{md_table}
\title{Build a multiple-decrement table}
\usage{
md_table(x, qxj, radix = 1e+05)
}
\arguments{
\item{x}{Integer vector of ages or durations.}

\item{qxj}{Matrix/data.frame of cause-specific decrement probabilities.
Rows correspond to ages in x, columns correspond to causes.}

\item{radix}{Starting \eqn{l_x^{(\tau)}}.}
}
\value{
Data frame containing \eqn{q^{(j)}}, \eqn{q^{(\tau)}},
  \eqn{p^{(\tau)}}, \eqn{l^{(\tau)}}, \eqn{d^{(j)}}, and \eqn{d^{(\tau)}}.
}
\description{
Build a multiple-decrement table
}
\examples{
x <- 45:50
qmat <- cbind(
  q1 = c(.011, .012, .013, .014, .015, .016),
  q2 = c(.100, .100, .100, .100, .100, .100)
)
md_table(x, qmat, radix = 1000)
}
