% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqriskR-package.R
\docType{package}
\name{mqriskR-package}
\alias{mqriskR}
\alias{mqriskR-package}
\title{mqriskR package}
\description{
Provides functions for actuarial risk modeling, including survival models, life annuities, multiple-decrement models, and mortality improvement projections. The package is designed to align with standard actuarial notation and supports teaching, exam preparation, and reproducible actuarial analysis. The methods are based on standard actuarial references including Camilli, Duncan and London (2014, ISBN:9781625423474) "Models for Quantifying Risk" and Dickson, Hardy and Waters (2020, ISBN:9781108478083) "Actuarial Mathematics for Life Contingent Risks".
}
\author{
\strong{Maintainer}: Nii Okine \email{okinean@appstate.edu}

Authors:
\itemize{
  \item Nii Okine \email{okinean@appstate.edu}
}

}
\keyword{internal}
