% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{multilife_ch12_ext}
\alias{multilife_ch12_ext}
\title{Extended multi-life functions for Chapter 12}
\description{
Additional Chapter 12 functions for contingent probabilities,
continuous contingent insurances, and continuous multi-life annuities.
}
\keyword{internal}
