% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{tpxy}
\alias{tpxy}
\alias{tqxy}
\alias{tpxybar}
\alias{tqxybar}
\title{Multi-life survival and failure probabilities}
\usage{
tpxy(x, y, t, tbl = NULL, model = NULL, ...)

tqxy(x, y, t, tbl = NULL, model = NULL, ...)

tpxybar(x, y, t, tbl = NULL, model = NULL, ...)

tqxybar(x, y, t, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of the first life. May be scalar or vector.}

\item{y}{Age of the second life. May be scalar or vector.}

\item{t}{Nonnegative duration. May be scalar or vector.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model.}

\item{...}{Additional parameters passed to the survival model.}
}
\value{
A numeric vector of probabilities.
}
\description{
Computes joint-life and last-survivor survival and failure probabilities
for two independent lives.
}
\details{
\code{tpxy()} computes
\eqn{{}_t p_{xy} = {}_t p_x\,{}_t p_y}.

\code{tqxy()} computes
\eqn{{}_t q_{xy} = 1 - {}_t p_{xy}}.

\code{tpxybar()} computes
\eqn{{}_t p_{\overline{xy}}
= {}_t p_x + {}_t p_y - {}_t p_{xy}}.

\code{tqxybar()} computes
\eqn{{}_t q_{\overline{xy}}
= 1 - {}_t p_{\overline{xy}}}.
}
\examples{
tpxy(40, 50, t = 10, model = "uniform", omega = 100)
tqxy(40, 50, t = 10, model = "uniform", omega = 100)
tpxybar(40, 50, t = 10, model = "uniform", omega = 100)
tqxybar(40, 50, t = 10, model = "uniform", omega = 100)

}
