% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{nAbarx}
\alias{nAbarx}
\title{Continuous deferred insurance APV}
\usage{
nAbarx(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Deferral period.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of APVs.
}
\description{
Computes
\eqn{{}_{n\mid}\bar{A}_x = v^n\,{}_np_x\,\bar{A}_{x+n}}.
}
