% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{nAbarx_udd}
\alias{nAbarx_udd}
\title{UDD approximation of continuous deferred insurance}
\usage{
nAbarx_udd(nAx, i)
}
\arguments{
\item{nAx}{Discrete deferred insurance APV.}

\item{i}{Effective annual interest rate.}
}
\value{
Continuous deferred insurance APV under UDD.
}
\description{
Computes \eqn{{}_{n\mid}\bar{A}_x = (i/\delta)\,{}_{n\mid}A_x}.
}
