% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{nAx_m_udd}
\alias{nAx_m_udd}
\title{UDD approximation of m-thly deferred insurance}
\usage{
nAx_m_udd(nAx, i, m)
}
\arguments{
\item{nAx}{Discrete deferred insurance APV.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}
}
\value{
m-thly deferred insurance APV under UDD.
}
\description{
Computes \eqn{{}_{n\mid}A_x^{(m)} = (i/i^{(m)})\,{}_{n\mid}A_x}.
}
