% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{NC_PUC_db}
\alias{NC_PUC_db}
\title{Projected Unit Credit normal cost}
\usage{
NC_PUC_db(projected_benefit, total_service, v_to_ret, p_surv, adue_ret)
}
\arguments{
\item{projected_benefit}{Nonnegative projected annual benefit at retirement.}

\item{total_service}{Positive total service at retirement.}

\item{v_to_ret}{Nonnegative discount factor from the valuation date to
retirement.}

\item{p_surv}{Survival or active-service probability to retirement in
\code{[0, 1]}.}

\item{adue_ret}{Positive retirement annuity-due factor.}
}
\value{
A numeric vector.
}
\description{
Computes the actuarial present value of the portion of the projected
benefit attributed to the current year of service.
}
\examples{
NC_PUC_db(
  projected_benefit = 30000,
  total_service = 30,
  v_to_ret = 0.5,
  p_surv = 0.9,
  adue_ret = 12
)

}
