% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_fractional.R
\name{nkqx}
\alias{nkqx}
\title{Curtate death probability {}_{k|} q_x from a life table}
\usage{
nkqx(tbl, x, k)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of ages.}

\item{k}{Nonnegative integer.}
}
\value{
Numeric vector of \eqn{{}_{k|} q_x} values.
}
\description{
Computes
\eqn{{}_{k|} q_x = {}_k p_x - {}_{k+1} p_x}
}
