% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_fractional.R
\name{nmxq}
\alias{nmxq}
\title{Deferred probability {}_{n|m} q_x from a life table}
\usage{
nmxq(tbl, x, n, m)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of ages.}

\item{n}{Nonnegative integer deferred period.}

\item{m}{Nonnegative integer subsequent period.}
}
\value{
Numeric vector of \eqn{{}_{n|m} q_x} values.
}
\description{
Computes the probability that a life aged x survives n years and then
dies within the following m years:
\eqn{{}_{n|m} q_x = {}_n p_x \cdot {}_m q_{x+n}}
}
\details{
This function is for integer n and m in the discrete tabular setting.
}
