% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_select.R
\name{nmxq_select}
\alias{nmxq_select}
\title{Deferred select-life death probability}
\usage{
nmxq_select(tbl, x_sel, t, n, m)
}
\arguments{
\item{tbl}{A select_life_table object.}

\item{x_sel}{Numeric vector of ages at selection.}

\item{t}{Numeric vector of current durations since selection.}

\item{n}{Numeric vector of nonnegative integer deferred periods.}

\item{m}{Numeric vector of nonnegative integer death windows.}
}
\value{
Numeric vector of deferred death probabilities.
}
\description{
Computes
\eqn{{}_{n|m} q_{[x]+t} = {}_n p_{[x]+t} \cdot {}_m q_{[x]+t+n}}.
}
