% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{NPV_profit}
\alias{NPV_profit}
\title{Net present value of a profit signature}
\usage{
NPV_profit(Pi, r)
}
\arguments{
\item{Pi}{Numeric profit-signature vector.}

\item{r}{Annual effective risk discount rate. May be scalar or vector;
values must be greater than \code{-1}.}
}
\value{
A numeric vector with one value for each rate in \code{r}.
}
\description{
Discounts a profit signature at one or more annual effective risk discount
rates.
}
\examples{
Pi <- c(-15.00, 8.42, 8.39, 8.58)
NPV_profit(Pi, r = 0.10)
NPV_profit(Pi, r = c(0.08, 0.10, 0.12))

}
