% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{npx}
\alias{npx}
\title{Compute n-year survival probability from a life table}
\usage{
npx(tbl, x, n)
}
\arguments{
\item{tbl}{A \code{life_table} object.}

\item{x}{Ages.}

\item{n}{Nonnegative integer durations.}
}
\value{
Numeric vector of \eqn{{}_n p_x} values.
}
\description{
Compute n-year survival probability from a life table
}
