% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_select.R
\name{npx_select}
\alias{npx_select}
\title{Select-life survival probability}
\usage{
npx_select(tbl, x_sel, t, n)
}
\arguments{
\item{tbl}{A select_life_table object.}

\item{x_sel}{Numeric vector of ages at selection.}

\item{t}{Numeric vector of current durations since selection.}

\item{n}{Numeric vector of nonnegative integer future durations.}
}
\value{
Numeric vector of survival probabilities.
}
\description{
Computes
\eqn{{}_n p_{[x]+t} = l_{[x]+t+n} / l_{[x]+t}}
}
\details{
in the discrete select-table setting.
}
