% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{nqx}
\alias{nqx}
\title{Compute {}_n q_x = {}_n d_x / l_x = 1 - {}_n p_x}
\usage{
nqx(tbl, x, n)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}

\item{n}{Nonnegative integer durations.}
}
\value{
Numeric vector of {}_n q_x values.
}
\description{
Compute {}_n q_x = {}_n d_x / l_x = 1 - {}_n p_x
}
