% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_select.R
\name{nqx_select}
\alias{nqx_select}
\title{Select-life death probability}
\usage{
nqx_select(tbl, x_sel, t, n)
}
\arguments{
\item{tbl}{A select_life_table object.}

\item{x_sel}{Numeric vector of ages at selection.}

\item{t}{Numeric vector of current durations since selection.}

\item{n}{Numeric vector of nonnegative integer future durations.}
}
\value{
Numeric vector of death probabilities.
}
\description{
Computes
\eqn{{}_n q_{[x]+t} = 1 - {}_n p_{[x]+t}}.
}
