% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_ch10_ext.R
\name{Pnax}
\alias{Pnax}
\title{Deferred annuity-immediate premium}
\usage{
Pnax(x, n, i, model, ...)
}
\arguments{
\item{x}{Issue age.}

\item{n}{Deferral period.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes
\eqn{P({}_{n|}a_x) = {}_{n|}a_x / \ddot{a}_{x:\angl{n}}}.
}
\examples{
Pnax(40, n = 20, i = 0.05, model = "uniform", omega = 100)
}
