% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{profit_margin}
\alias{profit_margin}
\title{Profit margin}
\usage{
profit_margin(NPV, APV_GP)
}
\arguments{
\item{NPV}{Net present value of profits.}

\item{APV_GP}{Positive actuarial present value of gross premiums.}
}
\value{
A numeric vector.
}
\description{
Computes net present value divided by the actuarial present value of gross
premiums. Scalar arguments are recycled to the common length.
}
\examples{
profit_margin(6.03, 259.52)

}
