% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_theory.R
\name{pv_cashflows}
\alias{pv_cashflows}
\title{Present value of cash flows at time 0}
\usage{
pv_cashflows(cf, t, i)
}
\arguments{
\item{cf}{Cash flow amounts (positive = inflow, negative = outflow).}

\item{t}{Times of cash flows (same length as cf).}

\item{i}{Effective interest rate.}
}
\value{
Present value at time 0.
}
\description{
Present value of cash flows at time 0
}
\examples{
pv_cashflows(c(-100, 60, 60), c(0, 1, 2), i = 0.10)
}
