% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{pxtau}
\alias{pxtau}
\title{Survival probability \eqn{p_x^{(\tau)}}}
\usage{
pxtau(qxj)
}
\arguments{
\item{qxj}{Numeric vector of cause-specific decrement probabilities.}
}
\value{
Numeric scalar.
}
\description{
Survival probability \eqn{p_x^{(\tau)}}
}
\examples{
pxtau(c(0.011, 0.100))
}
