% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter16_universal_life.R
\name{pxtau_ul}
\alias{pxtau_ul}
\title{One-year persistency rates for universal life}
\usage{
pxtau_ul(qd, qw, year_end_withdrawal = TRUE)
}
\arguments{
\item{qd}{Mortality probabilities.}

\item{qw}{Withdrawal probabilities.}

\item{year_end_withdrawal}{Logical; if \code{TRUE}, use
\eqn{(1-q^{(d)})(1-q^{(w)})}. Otherwise use \eqn{1-q^{(d)}-q^{(w)}}.}
}
\value{
Numeric vector.
}
\description{
Computes the one-year persistency rates \eqn{p_x^{(\tau)}} under either
Equation (16.14) or Equation (16.15).
}
\examples{
qd <- c(.001, .002, .003)
qw <- c(.02, .02, .03)
pxtau_ul(qd, qw)

}
