% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{qx_tab}
\alias{qx_tab}
\title{Compute one-year death probability from a life table}
\usage{
qx_tab(tbl, x)
}
\arguments{
\item{tbl}{A \code{life_table} object.}

\item{x}{Ages.}
}
\value{
Numeric vector of \eqn{q_x} values.
}
\description{
Compute one-year death probability from a life table
}
