% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter16_universal_life.R
\name{rt_ul}
\alias{rt_ul}
\title{Ratio \eqn{r_t = AV_t / GMF_t} capped at 1}
\usage{
rt_ul(AV, GMF)
}
\arguments{
\item{AV}{Account value.}

\item{GMF}{Guaranteed maturity fund.}
}
\value{
Numeric scalar.
}
\description{
Computes the ratio in Equation (16.16).
}
\examples{
rt_ul(AV = 4918.20, GMF = 14678.57)

}
