% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_models.R
\name{S0}
\alias{S0}
\title{Survival function for age-at-failure}
\usage{
S0(t, model, ...)
}
\arguments{
\item{t}{Numeric vector of times.}

\item{model}{One of \code{"uniform"}, \code{"exponential"},
\code{"gompertz"}, \code{"makeham"}, or \code{"weibull"}.}

\item{...}{Model parameters.}
}
\value{
Numeric vector of survival probabilities.
}
\description{
Computes \eqn{S_0(t)=Pr(T_0>t)}.
}
