% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_conversions.R
\name{S0_to_lx}
\alias{S0_to_lx}
\title{Convert survival probabilities to life-table values}
\usage{
S0_to_lx(S0, radix = 1e+05)
}
\arguments{
\item{S0}{Numeric vector of survival probabilities.}

\item{radix}{Positive radix \eqn{l_0}.}
}
\value{
Numeric vector of \eqn{l_x} values.
}
\description{
Converts Chapter 5 survival function values \eqn{S_0(x)} into
Chapter 6 life-table values \eqn{l_x = l_0 S_0(x)} using a chosen radix.
}
