% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tpx_tau_cf}
\alias{tpx_tau_cf}
\title{Total survival probability {}_t p_x^(tau) under constant forces}
\usage{
tpx_tau_cf(mu, t)
}
\arguments{
\item{mu}{Numeric vector of cause-specific forces.}

\item{t}{Time.}
}
\value{
Numeric scalar/vector.
}
\description{
Total survival probability {}_t p_x^(tau) under constant forces
}
\examples{
tpx_tau_cf(c(0.10, 0.20), 5)
}
