% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tpxprimej_cf}
\alias{tpxprimej_cf}
\title{Single-decrement survival under a constant force}
\usage{
tpxprimej_cf(mu, t)
}
\arguments{
\item{mu}{Nonnegative constant force of decrement. May be scalar or
vector.}

\item{t}{Nonnegative time. May be scalar or vector.}
}
\value{
A numeric vector.
}
\description{
Computes
\deqn{{}_tp_x^{\prime(j)}=\exp(-\mu_jt).}
}
\examples{
tpxprimej_cf(0.10, 5)
tpxprimej_cf(0.10, c(1, 5, 10))

}
