% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter16_universal_life.R
\name{tpxtau_ul}
\alias{tpxtau_ul}
\title{Cumulative persistency to the end of each policy year}
\usage{
tpxtau_ul(qd, qw, year_end_withdrawal = TRUE)
}
\arguments{
\item{qd}{Mortality probabilities.}

\item{qw}{Withdrawal probabilities.}

\item{year_end_withdrawal}{Logical; if \code{TRUE}, use Equation (16.15).}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{{}_tp_x^{(\tau)}} from the one-year persistency rates.
}
\examples{
qd <- c(.001, .002, .003)
qw <- c(.02, .02, .03)
tpxtau_ul(qd, qw)

}
