% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{tpxy}
\alias{tpxy}
\title{Joint-life survival probability}
\usage{
tpxy(x, y, t, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{t}{Time.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{{}_tp_{xy} = {}_tp_x\,{}_tp_y} under independence.
}
\examples{
tpxy(40, 50, t = 10, model = "uniform", omega = 100)
}
