% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{tpxybar}
\alias{tpxybar}
\title{Last-survivor survival probability}
\usage{
tpxybar(x, y, t, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{t}{Time.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{{}_tp_{\overline{xy}} = {}_tp_x + {}_tp_y - {}_tp_{xy}}.
}
\examples{
tpxybar(40, 50, t = 10, model = "uniform", omega = 100)
}
