% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_fractional.R
\name{tqx_tab}
\alias{tqx_tab}
\title{Fractional failure probability from a life table}
\usage{
tqx_tab(tbl, x, t, assumption = c("udd", "cf", "balducci"))
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Numeric vector of integer ages.}

\item{t}{Numeric vector of fractional durations with \eqn{0 \le t \le 1}.}

\item{assumption}{One of \code{"udd"}, \code{"cf"}, \code{"balducci"}.}
}
\value{
Numeric vector of \eqn{{}_t q_x} values.
}
\description{
Computes \eqn{{}_t q_x = 1 - {}_t p_x} for \eqn{0 \le t \le 1}.
}
