% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tqxj_cf}
\alias{tqxj_cf}
\title{Cause-specific probability {}_t q_x^(j) under constant forces}
\usage{
tqxj_cf(mu, j, t)
}
\arguments{
\item{mu}{Numeric vector of cause-specific forces.}

\item{j}{Cause index.}

\item{t}{Time.}
}
\value{
Numeric scalar/vector.
}
\description{
Cause-specific probability {}_t q_x^(j) under constant forces
}
\examples{
tqxj_cf(c(0.10, 0.20), j = 1, t = 5)
}
