% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tqxprimej_cf}
\alias{tqxprimej_cf}
\title{Single-decrement failure under a constant force}
\usage{
tqxprimej_cf(mu, t)
}
\arguments{
\item{mu}{Nonnegative constant force of decrement. May be scalar or
vector.}

\item{t}{Nonnegative time. May be scalar or vector.}
}
\value{
A numeric vector.
}
\description{
Computes
\deqn{{}_tq_x^{\prime(j)}
=1-\exp(-\mu_jt).}
}
\examples{
tqxprimej_cf(0.10, 5)

}
