% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12.R
\name{tqxy}
\alias{tqxy}
\title{Joint-life failure probability}
\usage{
tqxy(x, y, t, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{t}{Time.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{{}_tq_{xy} = 1 - {}_tp_{xy}}.
}
\examples{
tqxy(40, 50, t = 10, model = "uniform", omega = 100)
}
