% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{tqxy2}
\alias{tqxy2}
\title{Probability that (x) fails after (y) within n years}
\usage{
tqxy2(x, y, n, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{n}{Term in years.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{{}_n q_{xy}^{2} = {}_n q_x - {}_n q_{xy}^{1}}.
}
\examples{
tqxy2(40, 50, n = 10, model = "uniform", omega = 100)
}
