% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilife_ch12_ext.R
\name{tqyx2}
\alias{tqyx2}
\title{Probability that (y) fails after (x) within n years}
\usage{
tqyx2(x, y, n, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age of first life.}

\item{y}{Age of second life.}

\item{n}{Term in years.}

\item{tbl}{Life table.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes \eqn{{}_n q_{xy}^{\hspace{1mm}2} = {}_n q_y - {}_n q_{xy}^{\hspace{1mm}1}}.
}
\examples{
tqyx2(40, 50, n = 10, model = "uniform", omega = 100)
}
