% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_ch10_ext.R
\name{tVnax}
\alias{tVnax}
\title{Deferred annuity-immediate reserve}
\usage{
tVnax(x, n, t, i, model, ...)
}
\arguments{
\item{x}{Issue age.}

\item{n}{Deferral period.}

\item{t}{Duration.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes the reserve for an n-year deferred annuity-immediate
for \eqn{t < n}.
}
\examples{
tVnax(40, n = 20, t = 10, i = 0.05, model = "uniform", omega = 100)
}
