% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{var_Abarxn}
\alias{var_Abarxn}
\title{Variance of continuous endowment insurance PV}
\usage{
var_Abarxn(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of variances.
}
\description{
Variance of continuous endowment insurance PV
}
