% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{var_Abarxn1}
\alias{var_Abarxn1}
\title{Variance of continuous term insurance PV}
\usage{
var_Abarxn1(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of variances.
}
\description{
Computes
\eqn{\mathrm{Var}(\bar{Z}_{x:\angl{n}}^{1}) = {}^{2}\bar{A}_{x:\angl{n}}^{1} - (\bar{A}_{x:\angl{n}}^{1})^2}.
}
