% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{var_Axn1}
\alias{var_Axn1}
\title{Variance of term insurance PV}
\usage{
var_Axn1(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of variances.
}
\description{
Computes \eqn{\mathrm{Var}(Z_{x:\angl{n}}^{1}) = {}^{2}A_{x:\angl{n}}^{1} - (A_{x:\angl{n}}^{1})^2}.
}
