% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_mthly.R
\name{var_Axn1_m}
\alias{var_Axn1_m}
\title{Variance of m-thly term insurance PV}
\usage{
var_Axn1_m(x, n, i, m, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{m}{Positive integer payment frequency.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of variances.
}
\description{
Variance of m-thly term insurance PV
}
