% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_continuous.R
\name{var_nAbarx}
\alias{var_nAbarx}
\title{Variance of continuous deferred insurance PV}
\usage{
var_nAbarx(x, n, i, model, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Deferral period.}

\item{i}{Effective annual interest rate.}

\item{model}{Parametric survival model name.}

\item{...}{Additional model parameters passed to survival-model functions.}
}
\value{
Numeric vector of variances.
}
\description{
Variance of continuous deferred insurance PV
}
