% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{var_nAx}
\alias{var_nAx}
\title{Variance of deferred insurance PV}
\usage{
var_nAx(x, n, i, tbl = NULL, model = NULL, ..., tol = 1e-12, k_max = 5000)
}
\arguments{
\item{x}{Age.}

\item{n}{Deferral period.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}

\item{tol}{Numerical tolerance for truncating infinite sums.}

\item{k_max}{Maximum number of terms in the sum.}
}
\value{
Numeric vector of variances.
}
\description{
Variance of deferred insurance PV
}
