% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_discrete.R
\name{var_nEx}
\alias{var_nEx}
\title{Variance of pure endowment PV}
\usage{
var_nEx(x, n, i, tbl = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Age.}

\item{n}{Term.}

\item{i}{Effective annual interest rate.}

\item{tbl}{Optional life table object.}

\item{model}{Optional parametric survival model name.}

\item{...}{Additional arguments passed to survival-model functions.}
}
\value{
Numeric vector of variances.
}
\description{
Variance of pure endowment PV
}
