% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter16_universal_life.R
\name{Vprefloor_crvm_ul}
\alias{Vprefloor_crvm_ul}
\title{Pre-floor CRVM reserve for universal life}
\usage{
Vprefloor_crvm_ul(r, pvfb_minus_pvfp)
}
\arguments{
\item{r}{Ratio \eqn{r_t}.}

\item{pvfb_minus_pvfp}{Difference \eqn{(PVFB)_t - (PVFP)_t}.}
}
\value{
Numeric scalar.
}
\description{
Computes the pre-floor CRVM reserve from Equation (16.17).
}
\examples{
Vprefloor_crvm_ul(r = 0.33506, pvfb_minus_pvfp = 70)

}
