% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{AB_cae}
\alias{AB_cae}
\title{Accrued benefit for a career average earnings plan}
\usage{
AB_cae(salary_history, p)
}
\arguments{
\item{salary_history}{Numeric vector of annual salaries to date.}

\item{p}{Accrual percentage.}
}
\value{
Accrued benefit.
}
\description{
Computes the accrued benefit using actual salary history only.
}
\examples{
AB_cae(salary_history = c(100000, 104000, 108160), p = 1)
}
