% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_utils.R
\name{Abarxn1_udd}
\alias{Abarxn1_udd}
\title{UDD approximation of continuous term insurance}
\usage{
Abarxn1_udd(Axn1, i)
}
\arguments{
\item{Axn1}{Discrete term insurance APV.}

\item{i}{Effective annual interest rate.}
}
\value{
Continuous term insurance APV under UDD.
}
\description{
Computes \eqn{\bar{A}_{x:\angl{n}}^{1} = (i/\delta)A_{x:\angl{n}}^{1}}.
}
