% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{APV_NR_db}
\alias{APV_NR_db}
\title{APV of normal retirement benefit for a DB plan}
\usage{
APV_NR_db(PABz, v_to_ret, p_surv, adue_ret)
}
\arguments{
\item{PABz}{Projected annual benefit at retirement.}

\item{v_to_ret}{Discount factor from current age to retirement.}

\item{p_surv}{Active-service survival probability to retirement.}

\item{adue_ret}{Retirement annuity factor.}
}
\value{
Actuarial present value of the normal retirement benefit.
}
\description{
Computes Equation (18.10).
}
\examples{
APV_NR_db(PABz = 108008.66, v_to_ret = 1 / 1.06^30, p_surv = 0.8, adue_ret = 12)
}
