% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter16_universal_life.R
\name{coi_ul_typeB}
\alias{coi_ul_typeB}
\title{Cost of insurance for Type B universal life}
\usage{
coi_ul_typeB(B, qx, iq)
}
\arguments{
\item{B}{Face amount.}

\item{qx}{Mortality rate for the period.}

\item{iq}{Interest rate used in the cost-of-insurance calculation.}
}
\value{
Numeric vector.
}
\description{
Computes the one-period cost of insurance for a Type B universal life policy
using Equation (16.5a):
\deqn{
COI_t = \frac{B q_{x+t-1}}{1+i^q}.
}
}
\examples{
coi_ul_typeB(B = 100000, qx = 0.00076, iq = 0.03)

}
