% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_theory.R
\name{discount}
\alias{discount}
\title{Discount factor (1+i)^(-t)}
\usage{
discount(i, t)
}
\arguments{
\item{i}{Effective interest rate.}

\item{t}{Time (can be vector).}
}
\value{
Discount factor v^t.
}
\description{
Discount factor (1+i)^(-t)
}
\examples{
discount(0.05, 0:5)
}
