% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{dx}
\alias{dx}
\title{Compute d_x = l_x - l_{x+1}}
\usage{
dx(tbl, x)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}
}
\value{
Numeric vector of d_x values.
}
\description{
Compute d_x = l_x - l_{x+1}
}
