% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{dxtau}
\alias{dxtau}
\title{Total decrements \eqn{d_x^{(\tau)}}}
\usage{
dxtau(lxtau, qxj)
}
\arguments{
\item{lxtau}{Number alive at age x in the multiple-decrement table.}

\item{qxj}{Numeric vector of cause-specific decrement probabilities.}
}
\value{
Numeric scalar.
}
\description{
Total decrements \eqn{d_x^{(\tau)}}
}
\examples{
dxtau(1000, c(0.011, 0.100))
}
