% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{Income_dc}
\alias{Income_dc}
\title{Retirement income from a defined contribution accumulation}
\usage{
Income_dc(AVz, adue_z)
}
\arguments{
\item{AVz}{Accumulated value at retirement.}

\item{adue_z}{Whole life annuity-due factor at retirement age.}
}
\value{
Annual retirement income.
}
\description{
Converts a defined contribution accumulation to annual annuity-due income.
}
\examples{
Income_dc(AVz = 824211.35, adue_z = 12)
}
