% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_varying_benefits.R
\name{insurance_varying_benefits}
\alias{insurance_varying_benefits}
\title{Insurance models with varying benefits (Chapter 7)}
\description{
Functions for discrete and continuous contingent payment models with
varying benefits, matching Chapter 7 notation.
}
\details{
Included functions:
\itemize{
  \item \eqn{(IA)_x}
  \item \eqn{(IA)_{x:\angl{n}}^{1}}
  \item \eqn{(DA)_{x:\angl{n}}^{1}}
  \item \eqn{(\bar{I}\bar{A})_x}
  \item \eqn{(I\bar{A})_x}
  \item \eqn{(\bar{I}\bar{A})_{x:\angl{n}}^{1}}
  \item \eqn{(\bar{D}\bar{A})_{x:\angl{n}}^{1}}
  \item \eqn{(D\bar{A})_{x:\angl{n}}^{1}}
}
}
\keyword{internal}
