% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{md_table}
\alias{md_table}
\title{Construct a multiple-decrement table}
\usage{
md_table(x, qxj, radix = 1e+05)
}
\arguments{
\item{x}{Integer vector of consecutive ages or durations.}

\item{qxj}{Numeric matrix or data frame of cause-specific decrement
probabilities. Rows correspond to values in \code{x}, and columns
correspond to causes.}

\item{radix}{Initial value of \eqn{l_x^{(\tau)}}.}
}
\value{
An object of classes \code{"md_table"} and \code{"data.frame"}
  containing age or duration, cause-specific decrement probabilities,
  total decrement and survival probabilities, numbers alive, and
  cause-specific and total decrements.
}
\description{
Constructs a discrete multiple-decrement table from cause-specific
one-year decrement probabilities.
}
\examples{
ages <- 45:50
qmat <- cbind(
  withdrawal = c(0.011, 0.012, 0.013, 0.014, 0.015, 0.016),
  retirement = rep(0.100, 6)
)

md_table(ages, qmat, radix = 1000)

}
