% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqriskR-package.R
\docType{package}
\name{mqriskR-package}
\alias{mqriskR}
\alias{mqriskR-package}
\title{mqriskR package}
\description{
Provides functions for actuarial risk modeling, including survival models, life annuities, multiple-decrement models, and mortality improvement projections. The package is designed to align with standard actuarial notation and supports teaching, exam preparation, and reproducible actuarial analysis.
}
\author{
\strong{Maintainer}: Nii Okine \email{okinean@appstate.edu}

}
\keyword{internal}
