% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqriskR-package.R
\docType{package}
\name{mqriskR-package}
\alias{mqriskR}
\alias{mqriskR-package}
\title{mqriskR package}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
