% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqriskR-package.R
\docType{package}
\name{mqriskR-package}
\alias{mqriskR}
\alias{mqriskR-package}
\title{mqriskR: Models for Quantifying Risk}
\description{
Actuarial functions for survival models, life contingencies, reserves,
multiple-decrement models, variable interest, universal life, profit
analysis, and pension mathematics.
}
\author{
\strong{Maintainer}: Nii Okine \email{okinean@appstate.edu}

Authors:
\itemize{
  \item Nii Okine \email{okinean@appstate.edu}
}

}
\keyword{internal}
