% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{NC_PUC_db}
\alias{NC_PUC_db}
\title{Projected Unit Credit normal cost for a DB plan}
\usage{
NC_PUC_db(projected_benefit, total_service, v_to_ret, p_surv, adue_ret)
}
\arguments{
\item{projected_benefit}{Projected benefit at retirement.}

\item{total_service}{Total service at retirement.}

\item{v_to_ret}{Discount factor to retirement.}

\item{p_surv}{Active-service survival probability to retirement.}

\item{adue_ret}{Retirement annuity factor.}
}
\value{
PUC normal cost.
}
\description{
Computes the PUC normal cost as the APV of the portion of the projected
benefit attributed to the current year of service.
}
\examples{
NC_PUC_db(projected_benefit = 30000, total_service = 30, v_to_ret = 0.5, p_surv = 0.9, adue_ret = 12)
}
