% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter18_pensions.R
\name{NC_TUC_db}
\alias{NC_TUC_db}
\title{Traditional Unit Credit normal cost for a DB plan}
\usage{
NC_TUC_db(accrual_benefit, v_to_ret, p_surv, adue_ret)
}
\arguments{
\item{accrual_benefit}{Benefit accrued in the current year.}

\item{v_to_ret}{Discount factor to retirement.}

\item{p_surv}{Active-service survival probability to retirement.}

\item{adue_ret}{Retirement annuity factor.}
}
\value{
TUC normal cost.
}
\description{
Computes the TUC normal cost as the APV of the current year's accrual.
}
\examples{
NC_TUC_db(accrual_benefit = 1560, v_to_ret = 0.5, p_surv = 0.9, adue_ret = 12)
}
