% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{ndx}
\alias{ndx}
\title{Compute {}_n d_x = l_x - l_{x+n}}
\usage{
ndx(tbl, x, n)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}

\item{n}{Nonnegative integer durations.}
}
\value{
Numeric vector of {}_n d_x values.
}
\description{
Compute {}_n d_x = l_x - l_{x+n}
}
