% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{NPV_partial}
\alias{NPV_partial}
\title{Partial net present values}
\usage{
NPV_partial(Pi, r)
}
\arguments{
\item{Pi}{Profit signature vector.}

\item{r}{Risk discount rate.}
}
\value{
Numeric vector.
}
\description{
Computes the sequence
\eqn{NPV(0), NPV(1), \dots, NPV(n)}
of partial net present values from a profit signature.
}
\examples{
Pi <- c(-15.00, 8.42, 8.39, 8.58)
NPV_partial(Pi, r = 0.10)

}
