% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{NPV_profit}
\alias{NPV_profit}
\title{Net present value of a profit signature}
\usage{
NPV_profit(Pi, r)
}
\arguments{
\item{Pi}{Profit signature vector.}

\item{r}{Risk discount rate.}
}
\value{
Numeric scalar.
}
\description{
Computes the Chapter 17 net present value:
\deqn{
NPV = \sum_{t=0}^{n}\frac{\Pi_t}{(1+r)^t}.
}
}
\examples{
Pi <- c(-15.00, 8.42, 8.39, 8.58)
NPV_profit(Pi, r = 0.10)

}
