% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{npx}
\alias{npx}
\title{Compute {}_n p_x = l_{x+n} / l_x}
\usage{
npx(tbl, x, n)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}

\item{n}{Nonnegative integer durations.}
}
\value{
Numeric vector of {}_n p_x values.
}
\description{
Compute {}_n p_x = l_{x+n} / l_x
}
