% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{npxtau_md}
\alias{npxtau_md}
\title{n-year survival probability {}_n p_x^(tau) from a multiple-decrement table}
\usage{
npxtau_md(tbl, x, n)
}
\arguments{
\item{tbl}{Output from md_table().}

\item{x}{Starting age.}

\item{n}{Number of years.}
}
\value{
Numeric scalar.
}
\description{
n-year survival probability {}_n p_x^(tau) from a multiple-decrement table
}
\examples{
x <- 45:50
qmat <- cbind(q1 = c(.011, .012, .013, .014, .015, .016), q2 = rep(.1, 6))
tbl <- md_table(x, qmat, radix = 1000)
npxtau_md(tbl, x = 46, n = 3)
}
