% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_ch10_ext.R
\name{PnAdotx}
\alias{PnAdotx}
\title{Deferred annuity-due premium}
\usage{
PnAdotx(x, n, i, model, ...)
}
\arguments{
\item{x}{Issue age.}

\item{n}{Deferral period.}

\item{i}{Effective annual interest rate.}

\item{model}{Survival model.}

\item{...}{Additional model parameters.}
}
\value{
Numeric vector.
}
\description{
Computes
\eqn{P({}_{n|}\ddot{a}_x) = {}_{n|}\ddot{a}_x / \ddot{a}_{x:\angl{n}}}.
}
\examples{
PnAdotx(40, n = 20, i = 0.05, model = "uniform", omega = 100)
}
