% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter17_profit_analysis.R
\name{profit_margin}
\alias{profit_margin}
\title{Profit margin}
\usage{
profit_margin(NPV, APV_GP)
}
\arguments{
\item{NPV}{Net present value of profits.}

\item{APV_GP}{Actuarial present value of gross premiums.}
}
\value{
Numeric scalar.
}
\description{
Computes the Chapter 17 profit margin:
\deqn{
\text{Profit Margin} = \frac{NPV}{APV_{GP}}.
}
}
\examples{
profit_margin(6.03, 259.52)

}
