% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{pxtau}
\alias{pxtau}
\title{Total one-year survival probability}
\usage{
pxtau(qxj)
}
\arguments{
\item{qxj}{Numeric vector of cause-specific decrement probabilities.}
}
\value{
A numeric scalar.
}
\description{
Computes
\deqn{p_x^{(\tau)}=1-q_x^{(\tau)}.}
}
\examples{
pxtau(c(0.011, 0.100))

}
