% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{qx_dep_cf}
\alias{qx_dep_cf}
\title{Dependent probabilities \eqn{q_x^{(j)}} from independent probabilities \eqn{q_x^{\prime(j)}} under constant force}
\usage{
qx_dep_cf(qxprime)
}
\arguments{
\item{qxprime}{Numeric vector of independent probabilities.}
}
\value{
Numeric vector of dependent probabilities.
}
\description{
Dependent probabilities \eqn{q_x^{(j)}} from independent probabilities \eqn{q_x^{\prime(j)}} under constant force
}
\examples{
qx_dep_cf(c(0.20, 0.10))
}
