% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{qx_dep_sudd}
\alias{qx_dep_sudd}
\title{Dependent probabilities \eqn{q_x^{(j)}} from independent probabilities \eqn{q_x^{\prime(j)}} under SUDD}
\usage{
qx_dep_sudd(q1prime, q2prime)
}
\arguments{
\item{q1prime}{Independent probability for decrement 1.}

\item{q2prime}{Independent probability for decrement 2.}
}
\value{
Numeric vector \code{c(q1, q2)}.
}
\description{
Two-decrement case only.
}
\examples{
qx_dep_sudd(0.20, 0.10)
}
