% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table_core.R
\name{qx_tab}
\alias{qx_tab}
\title{Compute q_x = d_x / l_x}
\usage{
qx_tab(tbl, x)
}
\arguments{
\item{tbl}{A life_table object.}

\item{x}{Ages.}
}
\value{
Numeric vector of q_x values.
}
\description{
Compute q_x = d_x / l_x
}
