% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{qxprime_mudd}
\alias{qxprime_mudd}
\title{Independent probabilities q_x^{'(j)} from dependent probabilities q_x^(j) under MUDD}
\usage{
qxprime_mudd(qxj)
}
\arguments{
\item{qxj}{Numeric vector of dependent probabilities q_x^(j).}
}
\value{
Numeric vector of independent probabilities q_x^{'(j)}.
}
\description{
Independent probabilities q_x^{'(j)} from dependent probabilities q_x^(j) under MUDD
}
\examples{
qxprime_mudd(c(.20, .10))
}
