% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{qxprime_sudd}
\alias{qxprime_sudd}
\title{Independent probabilities q_x^{'(j)} from dependent probabilities q_x^(j) under SUDD}
\usage{
qxprime_sudd(q1, q2)
}
\arguments{
\item{q1}{Dependent probability for decrement 1.}

\item{q2}{Dependent probability for decrement 2.}
}
\value{
Numeric vector c(q1prime, q2prime).
}
\description{
Two-decrement case only.
}
\examples{
qxprime_sudd(0.20, 0.10)
}
