% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{qxtau}
\alias{qxtau}
\title{Total probability of decrement q_x^(tau)}
\usage{
qxtau(qxj)
}
\arguments{
\item{qxj}{Numeric vector of cause-specific decrement probabilities.}
}
\value{
Numeric scalar.
}
\description{
Total probability of decrement q_x^(tau)
}
\examples{
qxtau(c(0.011, 0.100))
}
