% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter16_universal_life.R
\name{rt_ul}
\alias{rt_ul}
\title{Account-value to guaranteed-fund ratio}
\usage{
rt_ul(AV, GMF)
}
\arguments{
\item{AV}{Nonnegative account value.}

\item{GMF}{Positive guaranteed maturity fund.}
}
\value{
A numeric vector.
}
\description{
Computes the ratio of account value to guaranteed maturity fund, capped at
one.
}
\examples{
rt_ul(AV = 4918.20, GMF = 14678.57)

}
