% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter14_multiple_decrement.R
\name{thiele_dVdt_01}
\alias{thiele_dVdt_01}
\title{Reserve derivatives for the disability model with recovery}
\usage{
thiele_dVdt_01(t, V0, V1, delta, Pbar, B, R, mu01, mu02, mu10, mu12)
}
\arguments{
\item{t}{Time.}

\item{V0}{Value of \eqn{{}_{t}\overline{V}^{(0)}}.}

\item{V1}{Value of \eqn{{}_{t}\overline{V}^{(1)}}.}

\item{delta}{Force of interest.}

\item{Pbar}{Continuous premium rate.}

\item{B}{Death benefit.}

\item{R}{Continuous disability income rate.}

\item{mu01}{Function of time returning \eqn{\mu_{x+t}^{01}}.}

\item{mu02}{Function of time returning \eqn{\mu_{x+t}^{02}}.}

\item{mu10}{Function of time returning \eqn{\mu_{x+t}^{10}}.}

\item{mu12}{Function of time returning \eqn{\mu_{x+t}^{12}}.}
}
\value{
A named numeric vector with components \code{dV0} and \code{dV1}.
}
\description{
Computes the right-hand sides of the coupled Thiele differential equations
in Equations (14.25) and (14.26) for the healthy-life reserve
\eqn{{}_{t}\overline{V}^{(0)}} and the disabled-life reserve
\eqn{{}_{t}\overline{V}^{(1)}}.
}
\details{
The equations are
\deqn{
\frac{d}{dt}{}_{t}\overline{V}^{(0)}
= \overline{P} + \delta {}_{t}\overline{V}^{(0)}
- \mu_{x+t}^{02}(B-{}_{t}\overline{V}^{(0)})
- \mu_{x+t}^{01}({}_{t}\overline{V}^{(1)}-{}_{t}\overline{V}^{(0)})
}
and
\deqn{
\frac{d}{dt}{}_{t}\overline{V}^{(1)}
= \delta {}_{t}\overline{V}^{(1)} - R
- \mu_{x+t}^{12}(B-{}_{t}\overline{V}^{(1)})
- \mu_{x+t}^{10}({}_{t}\overline{V}^{(0)}-{}_{t}\overline{V}^{(1)})
}
}
\examples{
mu01 <- function(t) 0.10 * t + 0.20
mu02 <- function(t) 0.20
mu10 <- function(t) 0.50
mu12 <- function(t) 0.125 * t + 0.20

thiele_dVdt_01(
  t = 2.0, V0 = 0, V1 = 0,
  delta = 0.04, Pbar = 446.95,
  B = 1000, R = 1000,
  mu01 = mu01, mu02 = mu02, mu10 = mu10, mu12 = mu12
)

}
