% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter14_multiple_decrement.R
\name{thiele_path_01}
\alias{thiele_path_01}
\title{Backward reserve path for the disability model with recovery}
\usage{
thiele_path_01(
  h,
  n,
  delta,
  Pbar,
  B,
  R,
  mu01,
  mu02,
  mu10,
  mu12,
  V0_n = 0,
  V1_n = 0
)
}
\arguments{
\item{h}{Step size.}

\item{n}{Final time.}

\item{delta}{Force of interest.}

\item{Pbar}{Continuous premium rate.}

\item{B}{Death benefit.}

\item{R}{Continuous disability income rate.}

\item{mu01}{Function of time returning \eqn{\mu_{x+t}^{01}}.}

\item{mu02}{Function of time returning \eqn{\mu_{x+t}^{02}}.}

\item{mu10}{Function of time returning \eqn{\mu_{x+t}^{10}}.}

\item{mu12}{Function of time returning \eqn{\mu_{x+t}^{12}}.}

\item{V0_n}{Terminal value of \eqn{{}_n\overline{V}^{(0)}}.}

\item{V1_n}{Terminal value of \eqn{{}_n\overline{V}^{(1)}}.}
}
\value{
A data frame with columns \code{t}, \code{tV0}, and \code{tV1}.
}
\description{
Computes the backward Euler reserve path for the healthy-life reserve
\eqn{{}_t\overline{V}^{(0)}} and disabled-life reserve
\eqn{{}_t\overline{V}^{(1)}} using Equations (14.27) and (14.28).
}
\examples{
mu01 <- function(t) 0.10 * t + 0.20
mu02 <- function(t) 0.20
mu10 <- function(t) 0.50
mu12 <- function(t) 0.125 * t + 0.20

thiele_path_01(
  h = 0.10, n = 2.0, delta = 0.04, Pbar = 446.95,
  B = 1000, R = 1000,
  mu01 = mu01, mu02 = mu02, mu10 = mu10, mu12 = mu12
)

}
