% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter14_multiple_decrement.R
\name{tp00_tp01_euler}
\alias{tp00_tp01_euler}
\title{Euler approximation for \eqn{{}_{t}p_{x}^{00}} and \eqn{{}_{t}p_{x}^{01}}}
\usage{
tp00_tp01_euler(h, n, mu01, mu02, mu10, mu12, p00_0 = 1, p01_0 = 0)
}
\arguments{
\item{h}{Step size.}

\item{n}{Final time.}

\item{mu01}{Function of time returning \eqn{\mu_{x+t}^{01}}.}

\item{mu02}{Function of time returning \eqn{\mu_{x+t}^{02}}.}

\item{mu10}{Function of time returning \eqn{\mu_{x+t}^{10}}.}

\item{mu12}{Function of time returning \eqn{\mu_{x+t}^{12}}.}

\item{p00_0}{Initial value of \eqn{{}_{0}p_{x}^{00}}.}

\item{p01_0}{Initial value of \eqn{{}_{0}p_{x}^{01}}.}
}
\value{
A data frame with columns \code{t}, \code{tp00}, \code{tp01},
  and \code{tp02}.
}
\description{
Computes the Euler approximations in the disability model allowing for
recovery, as in Equations (14.20) and (14.21).
}
\details{
The model uses three states:
\itemize{
  \item State 0: healthy
  \item State 1: disabled
  \item State 2: deceased
}
}
\examples{
mu01 <- function(t) 0.10 * t + 0.20
mu02 <- function(t) 0.20
mu10 <- function(t) 0.50
mu12 <- function(t) 0.125 * t + 0.20

tp00_tp01_euler(
  h = 0.10, n = 2.0,
  mu01 = mu01, mu02 = mu02, mu10 = mu10, mu12 = mu12
)

}
