% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tpxprimej_cf}
\alias{tpxprimej_cf}
\title{Single-decrement survival probability {}_t p_x^{'(j)} under constant force}
\usage{
tpxprimej_cf(mu, t)
}
\arguments{
\item{mu}{Force of decrement for cause j.}

\item{t}{Time.}
}
\value{
Numeric scalar/vector.
}
\description{
Single-decrement survival probability {}_t p_x^{'(j)} under constant force
}
\examples{
tpxprimej_cf(0.10, 5)
}
