% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tqxprime_mudd}
\alias{tqxprime_mudd}
\title{Independent probabilities {}_t q_x^{'(j)} under MUDD}
\usage{
tqxprime_mudd(qxj, t)
}
\arguments{
\item{qxj}{Numeric vector of dependent probabilities q_x^(j).}

\item{t}{Time in [0,1].}
}
\value{
Numeric vector.
}
\description{
Independent probabilities {}_t q_x^{'(j)} under MUDD
}
\examples{
tqxprime_mudd(c(.20, .10), t = 0.5)
}
