% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrement_ch13_ext.R
\name{tqxprimej_cf}
\alias{tqxprimej_cf}
\title{Single-decrement failure probability \eqn{{}_t q_x^{\prime(j)}} under constant force}
\usage{
tqxprimej_cf(mu, t)
}
\arguments{
\item{mu}{Force of decrement for cause j.}

\item{t}{Time.}
}
\value{
Numeric scalar/vector.
}
\description{
Single-decrement failure probability \eqn{{}_t q_x^{\prime(j)}} under constant force
}
\examples{
tqxprimej_cf(0.10, 5)
}
