% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter15_variable_interest.R
\name{z_from_coupon_annual}
\alias{z_from_coupon_annual}
\title{Bootstrap annual spot rates from annual coupon-bond yields}
\usage{
z_from_coupon_annual(maturity, coupon_yield, par = 1000)
}
\arguments{
\item{maturity}{Integer vector of maturities in years, in increasing order.}

\item{coupon_yield}{Numeric vector of annual coupon yields.}

\item{par}{Par value of each bond.}
}
\value{
Numeric vector of annual effective spot rates.
}
\description{
Bootstraps annual effective zero-coupon yields from par annual coupon-bearing
bond yields of the same maturities.
}
\examples{
maturity <- 1:4
coupon_yield <- c(0.02, 0.04, 0.06, 0.08)
z_from_coupon_annual(maturity, coupon_yield)

}
